# HolidaysApp-Backend/driving/api-rest/contracts/holidayapp-openapi-v3.yaml
openapi: 3.0.0
info: # Información general sobre la API
  title: HolidaysApp API Rest
  version: 1.0.0
  description: API Rest for HolidaysApp
servers: # URLs base del servidor
  - url: http://localhost:8080
    description: Generated server url
...
paths: # Todos los endpoints de la API
  /api-rest/organizations:
    get:
      tags:
        - Organizations
      summary: Get all organizations (optionally filtered by name)
      description: Get all organizations or filter by name using query parameter
      operationId: getAllOrganizations
      parameters:
        - name: name
          in: query
          description: Optional name to filter organizations
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationDTO'
        '400':
          description: Wrong Request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 400
                message: invalid parameters

        '401':
          description: Not authorized – invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 401
                message: No valid token provided

        '403':
          description: Denied – no permission to access the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 403
                message: You do not have permission to access this resource

        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 404
                message: The Organization was not found

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 500
                message: Unexpected error occurred
    post:
      tags:
        - Organizations
      summary: Create a new organization
      description: Creates a new organization and returns the created resource
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateDTO'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDTO'
        '400':
          description: Wrong Request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 400
                message: invalid input data

        '401':
          description: Not authorized – invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 401
                message: No valid token provided

        '403':
          description: Denied – no permission to access the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 403
                message: You do not have permission to access this resource

        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 404
                message: The Organization was not found

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 500
                message: Unexpected error occurred
    put:
      tags:
        - Organizations
      summary: Update an existing organization
      description: Updates an organization by its ID, which must be included in the request body
      operationId: updateOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationDTO'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDTO'
        '400':
          description: Wrong Request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 400
                message: invalid input data
        '401':
          description: Not authorized – invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 401
                message: No valid token provided

        '403':
          description: Denied – no permission to access the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 403
                message: You do not have permission to access this resource

        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 404
                message: The Organization was not found

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 500
                message: Unexpected error occurred
  /api-rest/organizations/{organizationId}:
    get:
      tags:
        - Organizations
      summary: Get Organization by ID
      description: Get a Organization of Organizations APIs by ID
      operationId: getOrganizationById
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDTO'
        '400':
          description: Wrong Request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 400
                message: invalid input data
        '401':
          description: Not authorized – invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 401
                message: No valid token provided

        '403':
          description: Denied – no permission to access the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 403
                message: You do not have permission to access this resource

        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 404
                message: The Organization was not found

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 500
                message: Unexpected error occurred
    delete:
      tags:
        - Organizations
      summary: Delete an organization
      description: Deletes an organization by its ID
      operationId: deleteOrganizationById
      parameters:
        - name: organizationId
          in: path
          required: true
          description: ID of the organization to delete
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationDTO'
        '400':
          description: Wrong Request – invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 400
                message: invalid input data
        '401':
          description: Not authorized – invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 401
                message: No valid token provided

        '403':
          description: Denied – no permission to access the resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 403
                message: You do not have permission to access this resource

        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 404
                message: The Organization was not found

        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2025-05-20T19:00:00Z'
                status: 500
                message: Unexpected error occurred
...
components:
  schemas:
    OrganizationDTO:
      description: This is the Organization model
      type: object
      required:
        - name
      properties:
        id:
          type: integer
        name:
          type: string
    OrganizationCreateDTO:
      description: This is the Organization model to create
      type: object
      required:
        - name
      properties:
        name:
          type: string
          x-field-extra-annotation: "@NotBlank\n@Size(min=3, max=20)"
          minLength: 3
          maxLength: 20
...
    ErrorResponse:
      description: Standard error model for API responses
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        message:
          type: string