# HolidaysApp-Backend/driving/api-rest/src/main/java/com/rob/driving/controllers/adapters/error/ErrorResponse.java
package com.rob.driving.controllers.adapters.error;

public class ErrorResponse {
    private final String timestamp;
    private final int status;
    private final String message;

    public ErrorResponse(int status, String message) {
        this.timestamp = java.time.ZonedDateTime.now().toString();
        this.status = status;
        this.message = message;
    }

    public String getTimestamp() {
        return timestamp;
    }

    public int getStatus() {
        return status;
    }

    public String getMessage() {
        return message;
    }
}